Finished work on DI. Made some of the methods in PhoneService and BrandService functional. Removed all tests files. Most were empty, some methods are no longer used (so don't need to be tested), and no functional tests existed anyway.

+++ WebstorePhones.Domain/Interfaces/IRepository.cs
@@ -14,7 +14,7 @@ namespace WebstorePhones.Domain.Interfaces
         /// </summary>
         /// <param name="id"></param>
         /// <returns>T.</returns>
-        T GetById(int id);
+        T GetById(long id);
         /// <summary>
         /// Create a T.
         /// </summary>
@@ -25,6 +25,6 @@ namespace WebstorePhones.Domain.Interfaces
         /// Delete by id.
         /// </summary>
         /// <param name="id"></param>
-        void Delete(int id);
+        void Delete(long id);
     }
 }
